requirements.txt
# Marina's Spectra - Autonomous Browser System
# Requirements for all components and dependencies

# Core Browser Automation
playwright>=1.40.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0

# Async Programming & Utilities
asyncio-throttle>=1.0.2
aiofiles>=23.2.1
aiohttp>=3.9.1
uvloop>=0.19.0; sys_platform != "win32"

# Core Dependencies (Required)
nltk>=3.8.1
Pillow>=10.1.0
numpy>=1.24.3

# Optional Natural Language Processing (ML Features)
# Install with: pip install Marina[ml]
# spacy>=3.7.2
# transformers>=4.35.0
# torch>=2.1.0
# sentence-transformers>=2.2.2

# Optional Computer Vision & Media Processing
# Install with: pip install Marina[cv]
# opencv-python>=4.8.1.78
# scikit-image>=0.22.0
# imagehash>=4.3.1

# Optional OCR & Text Recognition
# Install with: pip install Marina[ocr]
# easyocr>=1.7.0
# pytesseract>=0.3.10
# pdf2image>=1.16.3

# Optional Audio Processing
# Install with: pip install Marina[audio]
# whisper>=1.1.10
# librosa>=0.10.1
# soundfile>=0.12.1

# Database & Storage
sqlite3; python_version >= "3.0"
redis>=5.0.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Security & Encryption
cryptography>=41.0.7
keyring>=24.3.0
argon2-cffi>=23.1.0
bcrypt>=4.1.2

# Configuration & Serialization
PyYAML>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
marshmallow>=3.20.1

# HTTP & API
requests>=2.31.0
httpx>=0.25.2
urllib3>=2.1.0

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.6
py-cpuinfo>=9.0.0

# Error Tracking & Logging
sentry-sdk>=1.38.0
structlog>=23.2.0
colorlog>=6.8.0

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
pre-commit>=3.6.0

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0

# Data Analysis & Visualization
pandas>=2.1.4
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0

# Machine Learning Extensions (Optional)
# Install with: pip install Marina[ml]
# scikit-learn>=1.3.2
# joblib>=1.3.2
# xgboost>=2.0.2
# lightgbm>=4.1.0

# Face Recognition (Optional)
# Install with: pip install Marina[cv,face]
# face-recognition>=1.3.0
# dlib>=19.24.2

# Proxy & Network
pysocks>=1.7.1
stem>=1.8.2
requests-tor>=1.0.0

# Utility Libraries
tqdm>=4.66.1
click>=8.1.7
rich>=13.7.0
python-dateutil>=2.8.2
pytz>=2023.3
faker>=20.1.0

# Performance & Optimization
cachetools>=5.3.2
diskcache>=5.6.3
memory-profiler>=0.61.0
line-profiler>=4.1.1

# File Processing
python-magic>=0.4.27
chardet>=5.2.0
openpyxl>=3.1.2
python-docx>=1.1.0

# Web Scraping Utilities
beautifulsoup4>=4.12.2
lxml>=4.9.3
cssselect>=1.2.0
pyquery>=2.0.0

# Geographic & Location
geopy>=2.4.0
geoip2>=4.7.0
maxminddb>=2.2.0

# Email & Communication
smtplib; python_version >= "3.0"
imaplib; python_version >= "3.0"
email-validator>=2.1.0

# Time & Scheduling
schedule>=1.2.0
croniter>=2.0.1
python-crontab>=3.0.0

# Validation & Parsing
jsonschema>=4.20.0
cerberus>=1.3.5
python-slugify>=8.0.1
phonenumbers>=8.13.25

# Platform-specific dependencies
pywin32>=306; sys_platform == "win32"
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
python-xlib>=0.33; sys_platform == "linux"

# Optional AI/ML Models
# sentence-transformers[extra]>=2.2.2
# torch-audio>=2.1.0
# torchvision>=0.16.0
# huggingface-hub>=0.19.4

# Optional Database Drivers
# pymongo>=4.6.0
# cassandra-driver>=3.28.0
# neo4j>=5.14.1

# Optional Cloud Services
# boto3>=1.34.0
# google-cloud-storage>=2.10.0
# azure-storage-blob>=12.19.0

# Optional Message Queues
# celery>=5.3.4
# kombu>=5.3.4
# pika>=1.3.2

# Optional Monitoring Extensions
# grafana-client>=3.5.0
# influxdb-client>=1.38.0
# elasticsearch>=8.11.0

# Development Tools (Optional)
# jupyter>=1.0.0
# ipython>=8.18.1
# notebook>=7.0.6
# jupyterlab>=4.0.9
